move simulator_model to sim folder
